Roll tabnav helper functions into PrimerHelpers
